Move recipes and reformat tests
